pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tortured-phrase-classifier"
version = "0.1.0"
description = "A warrant-based, three-layer classification system for detecting tortured phrases as terminological violations in scientific literature"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "TPC Contributors" }
]
keywords = ["knowledge organization", "research integrity", "paper mills", "tortured phrases", "terminology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.10"
dependencies = [
    "pyahocorasick>=2.0.0",
    "sentence-transformers>=2.7.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0.1",
    "jsonschema>=4.22.0",
    "requests>=2.31.0",
    "pymupdf>=1.24.0",
    "nltk>=3.8.1",
    "tqdm>=4.66.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",
    "python-multipart>=0.0.9",
    "rich>=13.7.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
tpc = "tpc.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["tpc*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=tpc --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py310"
